configs/mvebu: new defconfig for Turris MOX
This adds the CZ.NIC Turris MOX Ethernet switch board, which uses a
Marvell Armada 3720 SoC (EBU family).

The defconfig builds a sdcard image which can be loaded directly by the
"distro boot" enabled U-Boot which is preinstalled in the SPI NOR flash
of the board. The boot.scr file is taken directly from the stock OpenWRT
system that the board is sold with.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>