board/lx2160acex7: add multiple kernel config fragment files
docker.config is generated from the output of
https://github.com/moby/moby/blob/master/contrib/check-config.sh.

The need for modules-builtin.config comes from the fact that the way in
which I deploy the kernel to the main NVME drive is by copying the
Buildroot generated kernel, but not modules. So I need all modules to be
changed to built-in.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>