board/lx2160ardb: new platform
Add Buildroot support for the NXP LX2160A-RDB (reference design board).
This is configured here to boot from the SD card. To my surprise, NXP
RCWs for LX2160A rev2 silicon don't have a pre-existing example for SD
boot, so until one is available, include the RCW through board files.

I would have been fine with the default U-Boot environment, except that
I would like to ship the MC firmware deployed as a file in the
filesystem, and that requires a custom 'mcinitcmd' environment variable
to load it. Thus, the 'env default -a' output was customized and
included here in its entirety.

Previous review feedback said that the U-Boot environment is too large:
https://lore.kernel.org/buildroot/2hhshlixqffg7r3o33klqzyvbxnpwllhhkskbi5igtewh6htlg@5huoxuxspqlv/
Short of me starting to strip away features, the best I can do is to
amortize the cost by taking away some variables and placing them in
other U-Boot environment variable files which can be reused.

- board/freescale/common/env-distro-boot-layerscape.txt: The U-Boot
  distro boot feature is generic, however for Layerscape it was
  customized through the introduction of the esbc_validate command, to
  support Secure Boot.
- board/freescale/common/env-dpaa2.txt: a reusable layer for other
  LS1088A, LS2088A, LX2160A based SoCs.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>
Reviewed-by: Vincent Jardin <vjardin@free.fr>